Added Authorization for authors to match ID's

+++ AlbertosMarket/Controllers/CommentController.cs
@@ -72,40 +72,8 @@ namespace AlbertosMarket.Controllers
             return View(comment);
         }
 
-        /*[Authorize(Roles = "Author")]
-        // GET: Comment/Create
-        public ActionResult Create(int? market_id)
-        {
-            ViewBag.AuthorID = new SelectList(db.Authors, "ID", "Name");
-            ViewBag.MarketID = new SelectList(db.Markets, "ID", "Title");
-            return View();
-        }
-
-        // POST: Comment/Create
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        [Authorize(Roles = "Author")]
-        public ActionResult Create([Bind(Include = "CommentID,MarketID,Content")] Comment comment)
-        {
-            if (ModelState.IsValid)
-            {
-                comment.CommentDate = DateTime.Now;
-                comment.AuthorID = User.Identity.GetUserId();
-                //comment.Author = db.Authors.Find(comment.AuthorID);
-                db.Comments.Add(comment);
-                db.SaveChanges();
-                // return RedirectToAction("Index");
-                return RedirectToAction("Details", "Market", new { id = comment.MarketID });
-            }
-
-            ViewBag.AuthorID = new SelectList(db.Authors, "ID", "Name", comment.AuthorID);
-            ViewBag.MarketID = new SelectList(db.Markets, "ID", "Title", comment.MarketID);
-            return View(comment);
-        }*/
-
         // GET: Comment/Edit/5
+        [Authorize(Roles = "Author")]
         public ActionResult Edit(int? id)
         {
             if (id == null)
@@ -117,6 +85,10 @@ namespace AlbertosMarket.Controllers
             {
                 return HttpNotFound();
             }
+            if (!comment.AuthorID.Equals(User.Identity.GetUserId()))
+            {
+                return View("~/Views/Shared/NotOwner.cshtml");
+            }
             ViewBag.AuthorID = new SelectList(db.Authors, "AuthorID", "Name", comment.AuthorID);
             ViewBag.MarketID = new SelectList(db.Markets, "ID", "Title", comment.MarketID);
             return View(comment);
@@ -127,13 +99,17 @@ namespace AlbertosMarket.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "CommentID,Content")] Comment comment)
+        public ActionResult Edit([Bind(Include = "AuthorID,CommentID,MarketID,CommentDate,Content")] Comment comment)
         {
+            if (!User.Identity.GetUserId().Equals(comment.AuthorID))
+            {
+                return View("~/Views/Shared/NotOwner.cshtml");
+            }
             if (ModelState.IsValid)
             {
                 db.Entry(comment).State = EntityState.Modified;
                 db.SaveChanges();
-                return RedirectToAction("Index");
+                return RedirectToAction("Details", "Market", new { id = comment.MarketID });
             }
             ViewBag.AuthorID = new SelectList(db.Authors, "AuthorID", "Name", comment.AuthorID);
             ViewBag.MarketID = new SelectList(db.Markets, "ID", "Title", comment.MarketID);
@@ -141,6 +117,7 @@ namespace AlbertosMarket.Controllers
         }
 
         // GET: Comment/Delete/5
+        [Authorize(Roles = "Author")]
         public ActionResult Delete(int? id)
         {
             if (id == null)
@@ -152,18 +129,27 @@ namespace AlbertosMarket.Controllers
             {
                 return HttpNotFound();
             }
+            if (!comment.AuthorID.Equals(User.Identity.GetUserId()))
+            {
+                return View("~/Views/Shared/NotOwner.cshtml");
+            }
             return View(comment);
         }
 
         // POST: Comment/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Author")]
         public ActionResult DeleteConfirmed(int id)
         {
             Comment comment = db.Comments.Find(id);
+            if (!comment.AuthorID.Equals(User.Identity.GetUserId()))
+            {
+                return View("~/Views/Shared/NotOwner.cshtml");
+            }
             db.Comments.Remove(comment);
             db.SaveChanges();
-            return RedirectToAction("Index");
+            return RedirectToAction("Details", "Market", new { id = comment.MarketID });
         }
 
         protected override void Dispose(bool disposing)
